Program refactored to work with server
Read and write methods in Program have been changed to work with a server, which is local for now. Small bug fix regarding read method, where error would occur when inputting a string as read amount. Changed name of the server api to ServerAPI from Program, to reflect its purpose, and to avoid duplicate file names.

Co-authored-by: Anton <avad@itu.dk>